Merger
This is basically everything Niksa and I managed to do when it comes to the frontend/backend merging. What's missing currently is: upload/export, roles (and therefore security), sessions (or something similar) should be generated once the user logs in, and logout which will destroy the session and finally some polishing after these are all done.